refactor(flows): move abgabe context and stringReplacements to shared location
