+++ app/flows/beratungshilfeFormular/context.ts
@@ -1,9 +1,9 @@
-import { beratungshilfeAbgabe } from "~/flows/beratungshilfeFormular/abgabe/context";
 import { beratungshilfeAnwaltlicheVertretung } from "./anwaltlicheVertretung/context";
 import { beratungshilfeFinanzielleAngaben } from "./finanzielleAngaben/context";
 import { beratungshilfeGrundvoraussetzungen } from "./grundvoraussetzung/context";
 import { beratungshilfePersoenlicheDaten } from "./persoenlicheDaten/context";
 import { beratungshilfeRechtsproblem } from "./rechtsproblem/context";
+import { abgabeContext } from "../shared/abgabe/context";
 
 export const beratungshilfeFormularContext = {
   ...beratungshilfeAnwaltlicheVertretung,
@@ -11,5 +11,5 @@ export const beratungshilfeFormularContext = {
   ...beratungshilfeRechtsproblem,
   ...beratungshilfeFinanzielleAngaben,
   ...beratungshilfePersoenlicheDaten,
-  ...beratungshilfeAbgabe,
+  ...abgabeContext,
 } as const;
