+++ app/flows/beratungshilfeFormular/stringReplacements.ts
@@ -2,10 +2,7 @@ import { findCourtIfUnique } from "~/services/gerichtsfinder/amtsgerichtData.ser
 import type { BeratungshilfeFormularContext } from ".";
 import { anwaltlicheVertretungDone } from "./anwaltlicheVertretung/guards";
 import { eigentumZusammenfassungDone } from "./finanzielleAngaben/eigentumZusammenfassungDone";
-import {
-  eigentumDone,
-  finanzielleAngabeGuards,
-} from "./finanzielleAngaben/guards";
+import { eigentumDone } from "./finanzielleAngaben/guards";
 import {
   andereUnterhaltszahlungenDone,
   ausgabenDone,
@@ -16,30 +13,6 @@ import {
 } from "./finanzielleAngaben/navStates";
 import { beratungshilfePersoenlicheDatenDone } from "./persoenlicheDaten/context";
 import { rechtsproblemDone } from "./rechtsproblem/context";
-
-export const getKinderStrings = (context: BeratungshilfeFormularContext) => {
-  const arrayIndex = context.pageData?.arrayIndexes.at(0);
-  if (
-    typeof arrayIndex === "undefined" ||
-    !context.kinder ||
-    arrayIndex > context.kinder.length + 1
-  )
-    return {};
-  if (arrayIndex < context.kinder.length)
-    return {
-      "kind#vorname": context.kinder?.[arrayIndex].vorname,
-      "kind#nachname": context.kinder?.[arrayIndex].nachname,
-    };
-};
-
-export const getArrayIndexStrings = (
-  context: BeratungshilfeFormularContext,
-) => {
-  const arrayIndex = context.pageData?.arrayIndexes.at(0);
-  return typeof arrayIndex !== "undefined"
-    ? { "array#index": String(arrayIndex + 1) }
-    : {};
-};
 
 export const getAmtsgerichtStrings = (
   context: BeratungshilfeFormularContext,
@@ -84,37 +57,6 @@ export const weiteresEinkommenStrings = (
   };
 };
 
-export const geldAnlagenStrings = (context: BeratungshilfeFormularContext) => {
-  return {
-    hasLebensversicherung:
-      context.hasGeldanlage === "yes" &&
-      context.geldanlagen?.some(
-        (geldanlage) => geldanlage.befristetArt === "lifeInsurance",
-      ),
-    hasBausparvertrag:
-      context.hasGeldanlage === "yes" &&
-      context.geldanlagen?.some(
-        (geldanlage) => geldanlage.befristetArt === "buildingSavingsContract",
-      ),
-    hasWertpapiere:
-      context.hasGeldanlage === "yes" &&
-      context.geldanlagen?.some(
-        (geldanlage) => geldanlage.art === "wertpapiere",
-      ),
-    hasGutenhabenKrypto:
-      context.hasGeldanlage === "yes" &&
-      context.geldanlagen?.some(
-        (geldanlage) => geldanlage.art === "guthabenkontoKrypto",
-      ),
-    hasGiroTagesSparkonto:
-      context.hasGeldanlage === "yes" &&
-      context.geldanlagen?.some(
-        (geldanlage) => geldanlage.art === "giroTagesgeldSparkonto",
-      ),
-    hasGrundeigentum: context.hasGrundeigentum === "yes",
-  };
-};
-
 export const ausgabenStrings = (context: BeratungshilfeFormularContext) => {
   return {
     hasSchwangerschaft:
@@ -135,20 +77,6 @@ export const ausgabenStrings = (context: BeratungshilfeFormularContext) => {
 
 export const getAnwaltStrings = (context: BeratungshilfeFormularContext) => {
   return { hasNoAnwalt: context.anwaltskanzlei !== "yes" };
-};
-
-export const eigentumZusammenfassungShowWarnings = (
-  context: BeratungshilfeFormularContext,
-) => {
-  return {
-    hasPartnerschaftYes: finanzielleAngabeGuards.hasPartnerschaftYes({
-      context,
-    }),
-    eigentumTotalWorthLessThan10000:
-      finanzielleAngabeGuards.eigentumTotalWorthLessThan10000({
-        context,
-      }),
-  };
 };
 
 export const getMissingInformationStrings = (
