--- app/flows/prozesskostenhilfeFormular/abgabe/context.ts
@@ -1,5 +1,0 @@
-import { z } from "zod";
-import { beratungshilfeAbgabe } from "~/flows/beratungshilfeFormular/abgabe/context";
-
-const _contextObject = z.object(beratungshilfeAbgabe).partial();
-export type ProzesskostenhilfeAbgabe = z.infer<typeof _contextObject>;
