+++ app/flows/shared/abgabe/context.ts
@@ -1,9 +1,9 @@
 import { z } from "zod";
 import { customRequiredErrorMessage } from "~/services/validation/YesNoAnswer";
 
-export const beratungshilfeAbgabe = {
+export const abgabeContext = {
   abgabeArt: z.enum(["online", "ausdrucken"], customRequiredErrorMessage),
 };
 
-const _contextObject = z.object(beratungshilfeAbgabe).partial();
-export type BeratungshilfeAbgabe = z.infer<typeof _contextObject>;
+const _contextObject = z.object(abgabeContext).partial();
+export type AbgabeContext = z.infer<typeof _contextObject>;
