+++ app/flows/shared/stringReplacements.ts
@@ -1,0 +1,76 @@
+import { BeratungshilfeFormularContext } from "../beratungshilfeFormular";
+import { finanzielleAngabeGuards } from "../beratungshilfeFormular/finanzielleAngaben/guards";
+import { ProzesskostenhilfeFormularContext } from "../prozesskostenhilfeFormular";
+
+export const getKinderStrings = (
+  context: BeratungshilfeFormularContext | ProzesskostenhilfeFormularContext,
+) => {
+  const arrayIndex = context.pageData?.arrayIndexes.at(0);
+  if (
+    typeof arrayIndex === "undefined" ||
+    !context.kinder ||
+    arrayIndex > context.kinder.length + 1
+  )
+    return {};
+  if (arrayIndex < context.kinder.length)
+    return {
+      "kind#vorname": context.kinder?.[arrayIndex].vorname,
+      "kind#nachname": context.kinder?.[arrayIndex].nachname,
+    };
+};
+
+export const getArrayIndexStrings = (
+  context: BeratungshilfeFormularContext | ProzesskostenhilfeFormularContext,
+) => {
+  const arrayIndex = context.pageData?.arrayIndexes.at(0);
+  return typeof arrayIndex !== "undefined"
+    ? { "array#index": String(arrayIndex + 1) }
+    : {};
+};
+
+export const geldAnlagenStrings = (
+  context: BeratungshilfeFormularContext | ProzesskostenhilfeFormularContext,
+) => {
+  return {
+    hasLebensversicherung:
+      context.hasGeldanlage === "yes" &&
+      context.geldanlagen?.some(
+        (geldanlage) => geldanlage.befristetArt === "lifeInsurance",
+      ),
+    hasBausparvertrag:
+      context.hasGeldanlage === "yes" &&
+      context.geldanlagen?.some(
+        (geldanlage) => geldanlage.befristetArt === "buildingSavingsContract",
+      ),
+    hasWertpapiere:
+      context.hasGeldanlage === "yes" &&
+      context.geldanlagen?.some(
+        (geldanlage) => geldanlage.art === "wertpapiere",
+      ),
+    hasGutenhabenKrypto:
+      context.hasGeldanlage === "yes" &&
+      context.geldanlagen?.some(
+        (geldanlage) => geldanlage.art === "guthabenkontoKrypto",
+      ),
+    hasGiroTagesSparkonto:
+      context.hasGeldanlage === "yes" &&
+      context.geldanlagen?.some(
+        (geldanlage) => geldanlage.art === "giroTagesgeldSparkonto",
+      ),
+    hasGrundeigentum: context.hasGrundeigentum === "yes",
+  };
+};
+
+export const eigentumZusammenfassungShowWarnings = (
+  context: BeratungshilfeFormularContext | ProzesskostenhilfeFormularContext,
+) => {
+  return {
+    hasPartnerschaftYes: finanzielleAngabeGuards.hasPartnerschaftYes({
+      context,
+    }),
+    eigentumTotalWorthLessThan10000:
+      finanzielleAngabeGuards.eigentumTotalWorthLessThan10000({
+        context,
+      }),
+  };
+};
